break up jb desktop root model more
