+++ products/jbrowse-desktop/src/rootModel/index.test.js
@@ -4,9 +4,9 @@
 import electron from 'electron'
 import PluginManager from '@jbrowse/core/PluginManager'
 import { getSnapshot } from 'mobx-state-tree'
-import corePlugins from './corePlugins'
-import rootModelFactory from './rootModel'
-jest.mock('./makeWorkerInstance', () => () => {})
+import corePlugins from '../corePlugins'
+import rootModelFactory from '.'
+jest.mock('../makeWorkerInstance', () => () => {})
 
 describe('Root MST model', () => {
   let rootModel
